Metal: Split MetalBackend in Forward.h and DeviceMTL
